requirements.txt
pyyaml>=3.1.0
docopt
schema
piexif
dateparser
progressbar2
